requirements.txt
# requirements.txt
gradio
python-docx
langchain
langchain-community
chromadb
sentence-transformers
ollama
numpy
faiss-cpu
tiktoken
torch
transformers